requirements.txt
openai-whisper
torch
torchaudio
ffmpeg-python
fastapi
uvicorn
celery
redis
python-multipart
boto3